+++ src/main/java/com/karrot/domain/user/UserController.java
@@ -4,7 +4,6 @@ import org.hibernate.cfg.NotYetImplementedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.TransactionSystemException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,8 +18,6 @@ import com.karrot.global.common.ResponseStatusEnum;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import javax.validation.ConstraintViolationException;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -65,48 +62,15 @@ public class UserController {
         HttpStatus resStatus;
 
         // process
-        try {
-            // request dto 객체를 service 레이어로 전달하여 처리한다
-            UsersResponseDTO resDto = userService.createUser(user);
+        UsersResponseDTO resDto = userService.createUser(user);
 
-            // response dto 객체로 ResponseDTOWrapper를 만들어 반환한다.
-            response = new ResponseDTOWrapper<>();
-            response.setData(List.of(resDto));
-            response.setResponseMessage("Create User Success");
-            response.setResponseStatus(ResponseStatusEnum.SUCCESS);
-            response.setOriginalStatus(HttpStatus.CREATED);
-            resStatus = HttpStatus.CREATED;
-        }
-        //TODO : [개선] 예외 처리 세분화, 400/403/409
-        catch (TransactionSystemException e) {
-            e.printStackTrace();
-            if(e.getCause() instanceof ConstraintViolationException){
-                response = new ResponseDTOWrapper<>();
-                response.setData(new ArrayList<UsersResponseDTO>());
-                response.setResponseMessage("Invalid input");
-                response.setResponseStatus(ResponseStatusEnum.FAIL_INTERNAL);
-                response.setOriginalStatus(HttpStatus.BAD_REQUEST); //400
-                resStatus = HttpStatus.BAD_REQUEST;
-            }
-            else {
-                response = new ResponseDTOWrapper<>();
-                response.setData(new ArrayList<UsersResponseDTO>());
-                response.setResponseMessage("Internal server error");
-                response.setResponseStatus(ResponseStatusEnum.FAIL_INTERNAL);
-                response.setOriginalStatus(HttpStatus.INTERNAL_SERVER_ERROR); //503
-                resStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            }
-
-        }
-        catch (Exception e) {
-            e.printStackTrace();
-            response = new ResponseDTOWrapper<>();
-            response.setData(new ArrayList<UsersResponseDTO>());
-            response.setResponseMessage("Internal server error");
-            response.setResponseStatus(ResponseStatusEnum.FAIL_INTERNAL);
-            response.setOriginalStatus(HttpStatus.INTERNAL_SERVER_ERROR); //503
-            resStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-        }
+        // response dto 객체로 ResponseDTOWrapper를 만들어 반환한다.
+        response = new ResponseDTOWrapper<>();
+        response.setData(List.of(resDto));
+        response.setResponseMessage("Create User Success");
+        response.setResponseStatus(ResponseStatusEnum.SUCCESS);
+        response.setOriginalStatus(HttpStatus.CREATED);
+        resStatus = HttpStatus.CREATED;
 
         // return
         return new ResponseEntity<>(response, resStatus);
